+++ db/migrate/20190820075040_create_group_endpoints.rb
@@ -1,0 +1,10 @@
+class CreateGroupEndpoints < ActiveRecord::Migration[5.1]
+  def change
+    create_table :group_endpoints do |t|
+      t.integer :group_id
+      t.bigint :endpoint_id
+
+      t.timestamps null: false
+    end
+  end
+end