+++ db/migrate/20190820080624_add_foreign_key_ref_on_group_endpoints.rb
@@ -1,0 +1,6 @@
+class AddForeignKeyRefOnGroupEndpoints < ActiveRecord::Migration[5.1]
+  def change
+    add_foreign_key :group_endpoints, :groups
+    add_foreign_key :group_endpoints, :endpoints
+  end
+end